draw batching
DrawItems now batch into a VertexArray by their RenderStates (i.e. matching states are batched to one call)

Minimap can now be zoomed in and out of

Classes inheriting from IAnchorable can be anchored to classes inheriting IAnchor

+++ JourneyCore.Lib/Game/Environment/Tiling/Tile.cs
@@ -10,7 +10,7 @@ namespace JourneyCore.Lib.Game.Environment.Tiling
     {
         public int Id { get; set; }
         public string Type { get; set; }
-        public float Probability { get; set; }
+        public double Probability { get; set; }
         public CustomProperty[] Properties { get; set; }
         public int Gid { get; set; }
         public IntRect TextureRect { get; set; }
